add support for numpy 2.0

+++ GTC/reporting.py
@@ -219,7 +219,7 @@ def k2_factor_sq(df=inf,p=95):
         
         # `fdtri` is the inverse of the cumulative F distribution
         # returning `x` such that `fdtr(dfn, dfd, x) = p`
-        return norm*special.fdtri(2.0,df-1.0,p)
+        return float(norm*special.fdtri(2.0,df-1.0,p))
         
     else:
         raise RuntimeError(f"invalid df={df!r}")
@@ -250,10 +250,10 @@ def k_factor(df=inf,p=95):
     
     if df > inf_dof:
         # inverse cumulative Gaussian distribution
-        return special.ndtri(p)
+        return float(special.ndtri(p))
     elif df >= 1:
         # inverse cumulative Student-t distribution
-        return special.stdtrit(df,p)
+        return float(special.stdtrit(df,p))
     else:
         raise RuntimeError( f"invalid df: {df}" )
    
@@ -284,7 +284,7 @@ def k_to_dof(k,p=95):
         p = (1.0 + p/100.0)/2.0         
         df = special.stdtridf(p,k) 
         
-        return df if df < inf_dof else inf 
+        return float(df) if df < inf_dof else inf
 
 #----------------------------------------------------------------------------
 def sensitivity(y,x):
